Move the accessor tests to their own directory
