+++ core/test/CMakeLists.txt
@@ -1,5 +1,6 @@
 include(${PROJECT_SOURCE_DIR}/cmake/create_test.cmake)
 
+add_subdirectory(accessor)
 add_subdirectory(base)
 add_subdirectory(factorization)
 add_subdirectory(log)
