+++ core/test/accessor/CMakeLists.txt
@@ -1,0 +1,4 @@
+ginkgo_create_test(index_span)
+ginkgo_create_test(range)
+ginkgo_create_test(reduced_row_major)
+ginkgo_create_test(scaled_reduced_row_major)
